# main/src/swagger.yaml
openapi: '3.0.0'
servers:
  - url: http://localhost:5804
info:
  version: '1.0.0'
  title: 'Corvina drive proxy api'
  description: Push/pull artifacts to corvina
paths:
...
  /artifact-registry/artifacts/search:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref:  "#/components/schemas/ArtifactSearch"
            example:
              page: 0
              pageSize: 100
              type:  "application/vnd.org.app.tar+gzip"
              isLatest: true
      responses:
        '200':
          description: Returns the list of matching artifacts
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/PageOfArtifacts"
  /artifact-registry/artifacts:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref:  "#/components/schemas/ArtifactPost"
            example:
              localPath: "%appdata%/org/projects/project1.mypro"
              repositoryName: "project1"
              version: "1.0.0"
              type:  "application/vnd.org.app.tar+gzip"
              labels:
                "key1": "value1" 
                "key2": "value2" 

      summary: Push an artifact to the registry repository from a local path 
      responses:
        '201':
          description: The artifact has been uploaded to repository
...
components:
  schemas:
...
    Artifact:
      description: Artificat structure
      type: object
      properties:
        id:
          type: string
        author:
          type: string
        size:
          type: number
        type:
          type: string
          description: MIME type
        uploadedAt:
          type: string
          description: ISO date
        labels:
          type: object
        repositoryName:
          type: string
        resourceUrl:
          type: string
          description: OCI url of the reosurce
        versions:
          type: array
          description: Version matching this content
          items:
            type: string
        isLatest:
          type: boolean
          description: is the latest uploaded version
    PageOfArtifacts:
      type: object
      properties:
        number:
          description: Current page
          type: number
        totalElements:
          description: Total number of elements
          type: number
        totalPages:
          description: Total number of pages
          type: number
        content:
          description: List of artifacts
          type: array 
          items:
            $ref: "#/components/schemas/Artifact"
    ArtifactPost:
      type: object
      required: [ "localPath", "repositoryName", "version", "type"]
      properties:
        localPath:
          description: Path of local file to store to corvina
          type: string
        repositoryName:
          description: Name of the repo
          type: string
        version:
          description: Tag version of the artifact push
          type: string
        type:
          description: File MIME type, for instance application/vnd.org.app.tar+gzip
          type: string
        labels:
          description: Optional JSON object of key=value pairs
          type: object
    ArtifactSearch:
      type: object
      required: 
        - page
        - pageSize
      properties:
        page:
          description: Pagination selected page
          type: number
        pageSize:
          description: Maximum number of entries returned from page
          type: number
        search:
          description: Free text search
          type: string
        type:
          description: Search only for a specify MIME type
          type: string
        labels:
          description: Search only for artifacts matching the givne key=value map
          type: object
        isLatest:
          description: Returns only the latest uploaded artifact
          type: boolean
...